+++ SceneKit_Neumorphism/SceneKit_Neumorphism/ContentView/SwiftUIView/ImportButton/ImportButton.swift
@@ -13,7 +13,7 @@ struct ImportButton: View
 {
     
     
-    @EnvironmentObject var environment: Environment
+    @EnvironmentObject var snapshot: Snapshot
     
     
     var name: String
@@ -37,20 +37,5 @@ struct ImportButton: View
             }
         }
         .buttonStyle(Capsule(name: self.name))
-        .background(
-            GeometryReader
-            {
-                buttonGeometry -> AnyView in
-
-                let background = Color.clear // Color.init(.sRGB, white: 0.0, opacity: 0.1)
-                    .onAppear()
-                    {
-                        self.environment.buttonFramesForNames[self.name] = buttonGeometry.frame(in: .named("ContentView"))
-                    }
-                ContentView.buttonFramesForNames[self.name] = buttonGeometry.frame(in: .named("ContentView"))
-                
-                return AnyView(background)
-            }
-        )
     }
 }